refactor: api와 page에서 사용하는 타입 용어 통일
