+++ src/pages/search/index/components/TripCards.tsx
@@ -1,4 +1,4 @@
-import { TripCard } from '@/common/apis/calendar/types';
+import { TripCard } from '@/common/apis/search/types';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { DEFAULT_IMAGE } from '../constants';
