+++ src/pages/search/index/hooks/useSearch.ts
@@ -1,10 +1,10 @@
-import { calendarApi } from '@/common/apis/calendar';
-import { SearchParams } from '@/common/apis/calendar/types';
+import { searchApi } from '@/common/apis/search';
+import { SearchParams } from '@/common/apis/search/types';
 import { useInfiniteQuery } from '@tanstack/react-query';
 
 const useSearch = ({ startDate, endDate, page = 0, size = 10 }: SearchParams) => {
   const getSearchItems = async ({ pageParam = 0 }: { pageParam: number }) => {
-    return await calendarApi.postSearchResult({ startDate, endDate, page: pageParam, size });
+    return await searchApi.postSearchResult({ startDate, endDate, page: pageParam, size });
   };
 
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isError, error, isLoading } = useInfiniteQuery({
